requirements.txt
requests
pyyaml

numpy
pandas

scikit-learn
joblib

matplotlib
seaborn
